requirements-update-html.txt
pandas>=1.3.0
google-cloud-bigquery>=2.30.0
google-cloud-storage>=2.0.0
gspread>=5.0.0
oauth2client>=4.1.3
db-dtypes>=1.0.0
numpy>=1.20.0
